requirements.txt
# UnityLink Bot Engine - Python Dependencies
# Install: pip install -r requirements.txt

# Web framework (connectors)
fastapi==0.115.0
uvicorn[standard]==0.30.0

# HTTP client (API calls to Fanvue, OF, OpenRouter)
httpx==0.27.0

# Database
supabase==2.28.0

# Cache/Queue
redis==5.0.0

# Telegram bot
python-telegram-bot[all]==22.6

# LLM client (OpenAI-compatible for Venice.ai / RunPod)
openai>=1.0.0

# Error tracking
sentry-sdk[fastapi]==2.0.0

# Environment
python-dotenv==1.0.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-httpx==0.30.0

# RAG memory embeddings (subscriber long-term memory)
# torch CPU-only (no NVIDIA CUDA — saves ~2GB in Docker images)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers>=3.0.0
